fixed loading of spider.fbx and enabled it in testing (fixed #23)
fixed a few IndexOutOfBounds exceptions that resulted from porting C code.
In c the texture arrays are allocated to a fixed size so adding a new texture is
done using an index. In kotlin we have to use add.

fixed another IOB-exception where the allocated size of a buffer was wrong

fixed material type detection (all comparison now done in lower case here)

enabled spider.fbx file for testing and adjusted the test so that it works
with both the spider.obj file and the spider.fbx file. The fbx version of
the test does not have different material info so we skip testing this for fbx.

Changed spider.fbx file to closer match spider.obj file (removed light and camera)